requirements.txt
requests
PyMuPDF
streamlit
sentence-transformers
faiss-cpu
openai
python-dotenv
langchain
nltk
langchain 
langchain-community
langchain_ollama